+++ tests/tests_resources/__init__.py
@@ -1,3 +1,5 @@
-import flexiznam
+from tests.tests_resources.mock_flexilims import MockFlexilims
+from .data_for_testing import populate_test_data, PROJECT_ID
 
-flexilims_session = flexiznam.get_flexilims_session(project_id="demo_project")
+flexilims_session = MockFlexilims(project_id=PROJECT_ID)
+populate_test_data(flexilims_session)